Remove shared methods
